Load and run real checkpoints from safetensors
TransformerLM reads a HuggingFace config and assembles the kernels to match
it, rather than hardcoding one graph the way CutileGPT hardcodes GPT-2's.
Weights come in through safetensors and DLPack, so bfloat16 survives.

Qwen3-0.6B, Qwen3-Reranker-0.6B, and Qwen2.5-0.5B-Instruct all agree with
transformers on argmax at every position, and generate the same text:

    "The capital of France is Paris. It is the largest city in Europe..."
    "def fibonacci(n):\n    if n == 0:\n        return 0\n    elif n == 1:..."

Logits differ by a fraction of a unit against |max| ~19, which is bfloat16
accumulating in a different order. Agreement is judged on argmax for that
reason: at lower ranks transformers itself rounds two candidates to the same
logit, so their ordering is a tie-break rather than a disagreement.

Three bugs, all of the same shape - a checkpoint field the loader did not read
would have produced fluent nonsense rather than an error:

- Qwen2 gives Q, K, and V a bias. Ignoring them ran fine and returned garbage
  ("The capital of France is);;\n订阅七 indemistributed"). The loader now
  applies them, and refuses outright on any tensor in the file it does not
  consume, so the next unread field is an error instead of bad output.
- RoPE base is not always top-level rope_theta. Qwen3 nests it under
  rope_parameters at 1e6, Glimmer at 500000 with a per-layer list whose zeros
  mean that layer is not rotated at all. Defaulting to 10000 would have run.
- Gemma 4 gives global layers their own head_dim (512 vs 256), KV head count,
  and RoPE base, and GLM-4-32B rotates only half of each head. Both are now
  refused by name rather than silently mis-run, which moves them out of the
  supported list in docs/MODELS.md.

Also fixed: the gated and GELU MLP kernels multiplied fp32 activations against
a bfloat16 weight tile, which mma rejects - they now meet the weight's dtype,
the way attention already did. RoPE used a tile sized to the sequence, so any
prompt whose length was not a power of two failed to compile; it uses a fixed
tile and relies on padded loads and clipped stores, verified at lengths 1, 5,
7, 100.

scripts/verify_model.py runs any local checkpoint against transformers.

+++ cutile_gpt/kernels/fused_mlp.py
@@ -98,7 +98,10 @@ def fused_mlp_kernel(
         w_proj_tile = ct.load(W_proj, index=(bid_n, h), shape=(TN, TK),
                              padding_mode=ct.PaddingMode.ZERO)
         # hidden_gelu: (TM, TK) @ w_proj_tile^T: (TK, TN) -> (TM, TN)
-        acc = ct.mma(hidden_gelu, ct.transpose(w_proj_tile), acc)
+        # GELU runs in fp32; mma needs matching operand dtypes, so the
+        # activations meet the weight tile rather than the other way round.
+        acc = ct.mma(hidden_gelu.astype(w_proj_tile.dtype),
+                     ct.transpose(w_proj_tile), acc)
 
     # Add bias for output projection
     b_proj_tile = ct.load(B_proj, index=(bid_n,), shape=(TN,),
